Separate code into different src files; removed warnings
